ensure 64 bit integers where needed
As some architectures use just 32 bits for ‘long’ type, we need to
change the type to ‘uint64_t’ to ensure the same behaviour on all
architectures. This fixes some failing tests on a MIPS machine.

+++ sha256.h
@@ -7,7 +7,7 @@
 struct sha256_state {
 	unsigned int h[8];
 	unsigned char buf[64];
-	unsigned long len;
+	unsigned long long len;
 };
 
 void sha256_init(struct sha256_state *state);
